+++ mobile/shared/components/geckoview/GeckoViewStartup.sys.mjs
@@ -219,6 +219,7 @@ export class GeckoViewStartup {
               "GeckoView:IPProtection:Deactivate",
               "GeckoView:IPProtection:Enroll",
               "GeckoView:IPProtection:RefreshUsage",
+              "GeckoView:IPProtection:ServerList:GetList",
             ],
           });