+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -20,6 +20,7 @@ exclude: true
 - Added  [`clearTrackingDb`][153.4] to ContentBlockingController to support clearing all data stored about blocked trackers in previous browsing sessions. ([bug 2042068]({{bugzilla}}2042068))
 - Added [`IPProtectionController.refreshUsage`][153.5] to request a refresh of proxy usage information, delivered asynchronously via [`Delegate#onUsageChanged`][153.6].
   ([bug 2042799]({{bugzilla}}2042799))
+- Added [`IPProtectionController.getServerList`][153.7] and [`IPProtectionController.Country`][153.8] to retrieve the list of countries available in the proxy serverlist.
 
 [153.1]: {{javadoc_uri}}/GeckoSession.ScrollPositionUpdate.SourceType.html
 [153.2]: {{javadoc_uri}}/GeckoSession.ScrollPositionUpdate.html#source
@@ -27,6 +28,8 @@ exclude: true
 [153.4]: {{javadoc_uri}}/ContentBlockingController.html#clearTrackingDb()
 [153.5]: {{javadoc_uri}}/IPProtectionController.html#refreshUsage()
 [153.6]: {{javadoc_uri}}/IPProtectionController.Delegate.html#onUsageChanged(org.mozilla.geckoview.IPProtectionController.UsageInfo)
+[153.7]: {{javadoc_uri}}/IPProtectionController.html#getServerList()
+[153.8]: {{javadoc_uri}}/IPProtectionController.Country.html
 
 ## v152
 - Added [`WebExtensionController.INSTALLATION_METHOD_RTAMO`][152.1], which should be used when a `WebExtension` is installed for the RTAMO feature. ([bug 2029607]({{bugzilla}}2029607))
@@ -1990,4 +1993,4 @@ to allow adding gecko profiler markers.
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport(android.content.Context,android.os.Bundle,java.lang.String)
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: 554a1ee06f39ab8cfd88132f24bf9b368480039a
+[api-version]: 33d26a0c254812faa00b9b5ba41055d25842f665